+++ R/grouped-df.r
@@ -14,7 +14,6 @@ grouped_df <- function(data, vars, drop = TRUE) {
   if (length(vars) == 0) {
     return(tbl_df(data))
   }
-
   assert_that(is.data.frame(data), is.list(vars), all(sapply(vars,is.name)), is.flag(drop))
   grouped_df_impl(data, unname(vars), drop)
 }
@@ -225,5 +224,3 @@ sample_group <- function(tbl, i, frac = FALSE, size, replace = TRUE,
 
   i[sample.int(n, size, replace = replace, prob = weight)]
 }
-
-