+++ R/join.r
@@ -102,6 +102,7 @@ common_by <- function(by = NULL, x, y) {
   if (is.list(by)) return(by)
 
   if (!is.null(by)) {
+    by <- by[!duplicated(by)]
     x <- names(by) %||% by
     y <- unname(by)