+++ inst/include/dplyr.h
@@ -44,6 +44,7 @@ using namespace Rcpp ;
 
 namespace dplyr {
     class LazySubsets ;
+    Symbol extract_column( SEXP, const Environment& ) ;
     Symbol get_column(SEXP, const Environment&, const LazySubsets& ) ;
     class Result ;
     class ResultSet ;