+++ src/api.cpp
@@ -2,7 +2,30 @@
 
 namespace dplyr{
 
-    Symbol get_column(SEXP arg, const Environment& env, const LazySubsets& subsets ){
+    DataFrameJoinVisitors::DataFrameJoinVisitors(const Rcpp::DataFrame& left_, const Rcpp::DataFrame& right_, Rcpp::CharacterVector names_left, Rcpp::CharacterVector names_right, bool warn_ ) :
+        left(left_), right(right_),
+        visitor_names_left(names_left),
+        visitor_names_right(names_right),
+        nvisitors(names_left.size()),
+        visitors(nvisitors),
+        warn(warn_)
+    {
+        std::string name_left, name_right ;
+        for( int i=0; i<nvisitors; i++){
+            name_left  = names_left[i] ;
+            name_right = names_right[i] ;
+
+            try{
+                visitors[i] = join_visitor( left[name_left], right[name_right], name_left, name_right, warn ) ;
+            } catch( const std::exception& ex ){
+                stop( "cannot join on columns '%s' x '%s': %s ", name_left, name_right, ex.what() ) ;
+            } catch( ... ){
+                stop( "cannot join on columns '%s' x '%s'", name_left, name_right ) ;
+            }
+        }
+    }
+
+    Symbol extract_column( SEXP arg, const Environment& env ){
       RObject value ;
       if( TYPEOF(arg) == LANGSXP && CAR(arg) == Rf_install("~") ){
         if( Rf_length(arg) != 2 || TYPEOF(CADR(arg)) != SYMSXP )
@@ -18,6 +41,11 @@ namespace dplyr{
         stop("column must return a single string") ;
       }
       Symbol res(STRING_ELT(value,0)) ;
+      return res ;
+    }
+
+    Symbol get_column(SEXP arg, const Environment& env, const LazySubsets& subsets ){
+      Symbol res = extract_column(arg, env) ;
       if( !subsets.count(res) ){
         stop("result of column() expands to a symbol that is not a variable from the data: %s", CHAR(PRINTNAME(res)) ) ;
       }
@@ -131,7 +159,7 @@ namespace dplyr{
 
                   break ;
                 }
-                if( CAR(head) == Rf_install("~")) return ;
+                if( CAR(head) == Rf_install("~")) break ;
                 if( CAR(head) == Rf_install("order_by") ) break ;
                 if( CAR(head) == Rf_install("function") ) break ;
                 if( CAR(head) == Rf_install("local") ) return ;