+++ src/bind.cpp
@@ -3,25 +3,54 @@
 using namespace Rcpp ;
 using namespace dplyr ;
 
+class DataFrameAbleVector {
+public:
+
+  DataFrameAbleVector() : data(){}
+
+  inline void push_back( SEXP x) {
+    data.push_back( DataFrameAble(x) ) ;
+  }
+
+  inline const DataFrameAble& operator[]( int i) const {
+    return data[i] ;
+  }
+
+  inline int size() const {
+    return data.size() ;
+  }
+
+  ~DataFrameAbleVector(){
+    while (data.size()) data.pop_back();
+  }
+
+private:
+  std::vector<DataFrameAble> data ;
+} ;
+
 template <typename Dots>
 List rbind__impl( Dots dots, SEXP id = R_NilValue ){
     int ndata = dots.size() ;
     int n = 0 ;
-    std::vector<DataFrameAble> chunks ;
+    DataFrameAbleVector chunks ;
     std::vector<int> df_nrows ;
 
+    int k=0 ;
     for( int i=0; i<ndata; i++) {
-      chunks.push_back( DataFrameAble( dots[i] ) ) ;
-
-      int nrows = chunks[i].nrows() ;
+      SEXP obj = dots[i] ;
+      if( Rf_isNull(obj) ) continue ;
+      chunks.push_back( obj ) ;
+      int nrows = chunks[k].nrows() ;
       df_nrows.push_back(nrows) ;
       n += nrows ;
+      k++ ;
     }
+    ndata = chunks.size() ;
     pointer_vector<Collecter> columns ;
 
     std::vector<String> names ;
-    int k=0 ;
 
+    k=0 ;
     Function enc2native( "enc2native" ) ;
     for( int i=0; i<ndata; i++){
         Rcpp::checkUserInterrupt() ;
@@ -132,17 +161,20 @@ template <typename Dots>
 List cbind__impl( Dots dots ){
   int n = dots.size() ;
 
-  std::vector<DataFrameAble> chunks ;
+  DataFrameAbleVector chunks ;
   for( int i=0; i<n; i++) {
-    chunks.push_back( DataFrameAble( dots[i] ) );
+    SEXP obj = dots[i] ;
+    if( !Rf_isNull(obj) )
+      chunks.push_back( dots[i] );
   }
+  n = chunks.size() ;
 
   // first check that the number of rows is the same
   const DataFrameAble& df = chunks[0] ;
   int nrows = df.nrows() ;
   int nv = df.size() ;
   for( int i=1; i<n; i++){
-    const DataFrameAble& current = dots[i] ;
+    const DataFrameAble& current = chunks[i] ;
     if( current.nrows() != nrows ){
       stop( "incompatible number of rows (%d, expecting %d)", current.nrows(), nrows ) ;
     }
@@ -157,7 +189,7 @@ List cbind__impl( Dots dots ){
   for( int i=0, k=0 ; i<n; i++){
       Rcpp::checkUserInterrupt() ;
 
-      const DataFrameAble& current = dots[i] ;
+      const DataFrameAble& current = chunks[i] ;
       CharacterVector current_names = current.names() ;
       int nc = current.size() ;
       for( int j=0; j<nc; j++, k++){